requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.9.2
django-environ==0.4.5
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.3
phonenumberslite==8.12.16
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
requests==2.25.1
soupsieve==2.1
sqlparse==0.3.1
urllib3==1.26.2
whitenoise==5.2.0
